New release: v0.2.2
Feature: Added more convenient option for the user to modify the photometry catalog files. The catalogs can now be copied to the SPEEDYFIT_MODELS directory and modified there by the user. Speedyfit will first look into that directory when looking for the catalogs files and only if it can't find them, it will use the catalogs shipped with the package.

Bugfix: Fixed issue where the user had to add a trailing '/' to the SPEEDYFIT_MODELS env var. Now os.path.join is used to create filepaths so there is no issue on whatever system it is used.

Docs: Some improvements in the documentation were made.